se realiza el endpoint para actualizar lso tipos de cambios, manejando transacciones

+++ MoneyExchange.Application.DTO/ExchangeTypeDto.cs
@@ -4,6 +4,6 @@
     {
         public string MonedaOrigen { get; set; }
         public string MonedaDestino { get; set; }
-        public decimal TipoCambio { get; set; }
+        public double TipoCambio { get; set; }
     }
 }
